+++ apps/ledger-box/src/modules/wallet-settings/wallet-settings-activity/index.ts
@@ -1,0 +1,2 @@
+export { WalletSettingsActivity } from '#/modules/wallet-settings/wallet-settings-activity/wallet-settings-activity';
+export { WalletSettingsActivityRoute } from '#/modules/wallet-settings/wallet-settings-activity/wallet-settings-activity-route';