+++ apps/ledger-box/src/modules/wallets/wallet-settings-danger-zone/index.ts
@@ -1,1 +1,0 @@
-export { WalletSettingsDangerZone } from '#/modules/wallets/wallet-settings-danger-zone/wallet-settings-danger-zone';