+++ apps/ledger-box/src/modules/wallets/wallet-settings-general/index.ts
@@ -1,1 +1,0 @@
-export { WalletSettingsGeneral } from '#/modules/wallets/wallet-settings-general/wallet-settings-general';